made a final screen that just thanks the user for taking the text_types in the future, I think it would be best if we could display the results there

+++ lib/widgets/text_types.dart
@@ -8,7 +8,7 @@ class SubTitle extends StatelessWidget {
   final String text;
   final TextAlign textAlign;
   final double minFontSize;
-  final double maxLines;
+  final int maxLines;
   final TextStyle? style;
 
   const SubTitle(
